refactor: extract region map game camera helpers into part file (#3878)
Move private camera clamp, zoom sync, viewport snapshot, and resize
handling from ct_region_map_game.dart into a dedicated part for Phase 3
flame map modularization without changing public API behavior.
